requirements.txt
requests~=2.28.1
python-telegram-bot
environs~=9.5.0